completed models and relationships
