+++ models/genre.py
@@ -4,8 +4,14 @@ Define the Review class for the 'reviews' table in the database.
 """
 from models import db
 from models.base import *
-from models.community import genre_communities
-from models.books import book_genres
+
+genre_communities = db.Table(
+    "genre_communities",
+    db.Column("genre_id", db.Integer, db.ForeignKey("genre.id"), primary_key=True),
+    db.Column(
+        "community_id", db.Integer, db.ForeignKey("community.id"), primary_key=True
+    ),
+)
 
 
 class Genre(Base, db.Model):
@@ -19,7 +25,9 @@ class Genre(Base, db.Model):
 
     __tablename__ = "genre"
     name = db.Column(db.String(60), nullable=False)
-    books = db.relationship("Book", secondary=book_genres, back_populates="genre")
+    books = db.relationship("Book", backref="genre")
     communities = db.relationship(
-        "Community", secondary=genre_communities, back_populates="genre"
+        "Community",
+        secondary=genre_communities,
+        backref=db.backref("genre", lazy="dynamic"),
     )
