+++ models/reviews.py
@@ -25,4 +25,4 @@ class Review(Base, db.Model):
     user_id = db.Column(db.String(60), db.ForeignKey("user.id"), nullable=False)
     book_id = db.Column(db.String(60), db.ForeignKey("book.id"), nullable=False)
     review_text = db.Column(db.String(200), nullable=False)
-    user = db.relationship('User', back_populates='review')
+    # user = db.relationship("User", back_populates="review")
